chore(icons): further fixes
